introduce TrackerQueryBuilder base to share the common stuff
